Encapsulating code more efficently/ seperating concerns. And making sure users balance reflects stripe fee
